requirements.txt
numpy==2.2.6
opencv-python==4.12.0.88
ultralytics==8.3.201
requests==2.32.5
aiortc==1.13.0
aiohttp==3.12.15
python-dotenv==1.1.1
